requirements.txt
python-telegram-bot==21.6
httpx==0.27.2
SQLAlchemy==2.0.36
python-dotenv==1.0.1
dateparser==1.2.0
pydantic==2.9.2
loguru==0.7.2
pytest==8.3.3
pytest-asyncio==0.24.0